Move Environment class declaration
